requirements.txt
python-socketio[client]~=4.1.0
